+++ packages/backend/src/modules/discovery/DiscoveryModule.ts
@@ -3,11 +3,8 @@ import { providers } from 'ethers'
 
 import { Config } from '../../config'
 import { ConfigReader } from '../../core/discovery/ConfigReader'
-import { discover } from '../../core/discovery/discover'
-import { DiscoveryOptions } from '../../core/discovery/DiscoveryOptions'
-import { ProviderWithCache } from '../../core/discovery/provider/ProviderWithCache'
-import { saveDiscoveryResult } from '../../core/discovery/saveDiscoveryResult'
 import { ApplicationModule } from '../ApplicationModule'
+import { runDiscovery } from '../../core/discovery/runDiscovery'
 
 export function createDiscoveryModule(
   config: Config,
@@ -35,36 +32,8 @@ export function createDiscoveryModule(
   const start = async () => {
     logger = logger.for('DiscoveryModule')
     logger.info('Starting')
-    const projectConfig = await configReader.readConfig(safeConfig.project)
-    const overrides = projectConfig.overrides ?? {}
-    // Temporary mapping from new config structure to the old one
-    const discoveryOptions: DiscoveryOptions = {
-      skipAddresses: [],
-      skipMethods: {},
-      addAbis: {},
-    }
-    Object.keys(overrides).forEach((address) => {
-      const ignoreMethods = overrides[address].ignoreMethods
-      if (ignoreMethods) {
-        discoveryOptions.skipMethods[address] = ignoreMethods
-      }
-    })
 
-    const blockNumber =
-      safeConfig.blockNumber ?? (await provider.getBlockNumber())
-
-    const discoveryProvider = new ProviderWithCache(
-      provider,
-      etherscanClient,
-      blockNumber,
-    )
-
-    const result = await discover(
-      discoveryProvider,
-      projectConfig.initialAddresses,
-      discoveryOptions,
-    )
-    await saveDiscoveryResult(result, safeConfig.project, blockNumber)
+    await runDiscovery(provider, etherscanClient, configReader, safeConfig)
   }
 
   return {
